refactor(web): 拆 Generate.tsx 到 generate/ 子目录（0 行为变化）
测试页面重构 commit 3/N —— 纯重构，给后续 InlineLoraPicker / XY UI /
Compare view commits 准备空间。所有组件按职责拆到独立文件，主页面只剩
GeneratePage 编排逻辑。

文件拆分：
  Generate.tsx (479 行) →
    generate/types.ts        — RecentLora 接口 + DEFAULT_NEG 常量
    generate/useRecentLoras.ts — 启动时拉所有项目最新 LoRA 的 hook
    generate/SampleGallery.tsx — sample 缩略图 + 大图预览
    generate/LoraList.tsx    — 多 LoRA 编辑列表（commit 4 会被替换）
    generate/PromptList.tsx  — 多 prompt 编辑列表
    generate/NumField.tsx    — 带 label 的数值输入框
    generate/StatusBadge.tsx — task 状态徽章
    Generate.tsx (~190 行)   — 仅留 GeneratePage 主组件 + 状态机 + SSE

行为完全相同：UI 渲染、enqueue payload、SSE 处理、所有交互不变。

测试：tsc --noEmit 0 errors；vitest run 59/59 passed（无新增测试，
重构本身不增功能）。
