+++ studio/web/src/pages/tools/generate/types.ts
@@ -1,0 +1,10 @@
+/** 测试页面（Generate）共用本地类型 + 常量。 */
+
+export interface RecentLora {
+  label: string
+  path: string
+  createdAt: number
+}
+
+export const DEFAULT_NEG =
+  'worst quality, low quality, score_1, score_2, score_3, blurry, jpeg artifacts, bad anatomy, bad hands, bad feet, missing fingers, extra fingers, text, watermark, logo, signature'
